LayoutTests/platform/ios-simulator/ios/fast/coordinates/element-client-rects-expected.txt

This tests Element.getBoundingClientRects and getClientRects positions when unscaled, scaled, and panned.
On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
unscaled
PASS clientRect.left is 100
PASS clientRect.right is 200
PASS clientRect.top is 100
PASS clientRect.bottom is 200
PASS clientRect.width is 100
PASS clientRect.height is 100
PASS clientRect.left is 100
PASS clientRect.right is 200
PASS clientRect.top is 100
PASS clientRect.bottom is 200
PASS clientRect.width is 100
PASS clientRect.height is 100

scaled and panned
PASS clientRect.left is 90
PASS clientRect.right is 190
PASS clientRect.top is 90
PASS clientRect.bottom is 190
PASS clientRect.width is 100
PASS clientRect.height is 100
PASS clientRect.left is 90
PASS clientRect.right is 190
PASS clientRect.top is 90
PASS clientRect.bottom is 190
PASS clientRect.width is 100
PASS clientRect.height is 100
PASS successfullyParsed is true

TEST COMPLETE
